adding some shortcuts

+++ main.py
@@ -4,10 +4,13 @@ from tkinter import filedialog, messagebox
 from tkinter import *
 import os
 import platform
+import tkinter.filedialog
+import tkinter.messagebox
 
 
 
-def open_file():
+
+def open_file(event=None):
     "Ouvrir un fichier à partir de l'onglet Fichier dans l'IDE"
     file_path = filedialog.askopenfilename(filetypes= [("Draw++ Files", "*.dpp")])
     print("%d", file_path) #file_path retourne le chemin exact du fichier selectionné
@@ -19,10 +22,18 @@ def open_file():
     else:
         text_area.insert(tkinter.END, "erreur lors de la récupération du fichier")
 
-def run():
+def run(event=None):
     #résultat de l'appui du bouton run : sauvegarder dans le dossier .to_run
-    file_path =  './.to_run/to_execute.dpp'
-    os.system("rm ./.to_run/*.dpp")
+    tkinter.messagebox.showinfo("Run","Exécution du Code Draw++ fait avec succès")
+    #on verifie le system d'exploitation sur lequel on est
+    if platform.system() == "Linux":
+        file_path =  './.to_run/to_execute.dpp'
+        os.system("rm ./.to_run/*.dpp")
+    elif platform.system() == "Windows":
+        tkinter.messagebox.showwarning("Attention","MyDrawpp IDE n'est pas encore compatible avec  " + platform.system())
+    else:
+        tkinter.messagebox.showerror("Erreur","MyDrawpp IDE ne fonctionne pas avec " + platform.system())
+    
     
     print("%d", file_path)
     if file_path: #si le fichier a bien ete enregistrer
@@ -31,6 +42,29 @@ def run():
         window.title(f"Éditeur de Code Draw++ - {os.path.basename(file_path)}")
    
     print("Running the code...")
+
+
+
+def save_as_file(event=None):
+#on verifie le system d'exploitation sur lequel on est
+    if platform.system() == "Linux":
+        if not os.path.isdir("MyDrawings"):
+            os.system("mkdir MyDrawings")
+        file_path =  tkinter.filedialog.asksaveasfilename(initialdir= "./MyDrawings",filetypes=[("Draw++ Files", "*.dpp")])
+    elif platform.system() == "Windows":
+        tkinter.messagebox.showwarning("Attention","MyDrawpp IDE n'est pas encore compatible avec  " + platform.system())
+    else:
+        tkinter.messagebox.showerror("Erreur","MyDrawpp IDE ne fonctionne pas avec " + platform.system())
+    
+    print("%d", file_path)
+    if file_path: #si le fichier a bien ete enregistrer
+        with open(file_path, "w") as file:
+            file.write(text_area.get(1.0, tkinter.END)) #je get absolument tout le contenu de ma text_area
+        window.title(f"Éditeur de Code Draw++ - {os.path.basename(file_path)}")
+
+
+
+
 
 
 # je creer le fichier .to_run pour executer les fichiers .dpp avec le compilateur a venir
@@ -57,6 +91,7 @@ window.config(menu=my_menu)
 file_menu = tkinter.Menu(my_menu, tearoff=False)
 my_menu.add_cascade(label="File", menu=file_menu)
 file_menu.add_command(label="Open File", command=open_file)
+file_menu.add_command(label="Save As", command=save_as_file)
 
 
 #Bouton run
@@ -68,6 +103,10 @@ my_menu.add_command(label="Run",background='green', command=run)
 text_area = tkinter.Text(window, wrap="char", undo=True, bg="#353434", fg="white", insertbackground="white", font="Droid", highlightcolor="grey")
 text_area.pack(fill=tkinter.BOTH, expand=1)
 
+#shortcuts raccourcis
+window.bind('<Control-Shift-S>', save_as_file)
+window.bind('<Control-o>', open_file)
+window.bind('<Alt-Return>', run)
 
 
 #lancement de la page
